refactor: 상수 추가, 파일 분리 및 적용
