+++ src/constants/regexs.js
@@ -1,0 +1,5 @@
+const REGEXS = Object.freeze({
+  number: /^[0-9]+$/,
+});
+
+export default REGEXS;
